requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.14
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.0.0
filelock==3.8.0
fonttools==4.37.3
gitdb==4.0.9
GitPython==3.1.27
identify==2.5.5
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
jupyterlab-widgets==3.0.3
kaggle==1.5.12
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
nodeenv==1.7.0
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.20.2
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.0
requests==2.28.1
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stack-data==0.5.0
text-unidecode==1.3
timm==0.6.7
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing-extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
wandb==0.13.3
wcwidth==0.2.5
widgetsnbextension==4.0.3
